backend/requirements.txt
google-adk
vertexai
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
fastapi
uvicorn[standard]
python-dotenv
google-auth
PyMuPDF
requests
beautifulsoup4
tavily-python
aiohttp
